+++ PreOrclBackEnd/PreOrclApi/Models/SisPerPersona.cs
@@ -12,7 +12,7 @@ namespace PreOrclApi.Models
     {
         [PrimaryKey(AutoIncrement = false)]
         [Key]
-        public int per_IDPER { get; set; }
+        public double per_IDPER { get; set; }
         public string per_nombre_razon { get; set; }
         public string per_apellido_comercial { get; set; }
         public string per_nit { get; set; }
